requirements.txt
click
colorama
PyYAML
Rich
